fix: apply maintainer review to the patch command
Address sferik's review on #1280:

- Diff `--merge-base <base>` instead of `<base>...HEAD` so uncommitted
  work counts: running `simplecov patch` before a commit now scores the
  lines just written, while CI (a clean HEAD checkout) is unchanged.
- Reject a stray positional argument. A forgotten `--base`
  (`simplecov patch feature`) used to be silently dropped and diff the
  default, passing a `--minimum` gate over the wrong change.
- Gate on an exact Rational minimum. `minimum * relevant` overshoots in
  binary float (`64.4 * 250` lands at 16_100.000…2), failing a patch at
  exactly the floor; `Rational(minimum.to_s) * relevant` is exact.
- Extract the git half into `Patch::ChangedLines` and the presentation
  and gate into `Patch::Output`, dropping `Patch` under RuboCop's
  ModuleLength limit without a disable.
